Removed warnings on 64bit platforms.
darcs-hash:20060313195238-65a35-458275dcc4e0781902437c82c474d5d9180ea0d2.gz

+++ libsofia-sip-ua/soa/soa.c
@@ -717,7 +717,7 @@ int soa_set_capability_sdp(soa_session_t *ss,
 {
   SU_DEBUG_9(("soa_set_capability_sdp(%s::%p, %p, %p, %d) called\n",
 	      ss ? ss->ss_actions->soa_name : "", ss, sdp, str, 
-	      str && len == -1 ? strlen(str) : len));
+	      str && len == -1 ? (int)strlen(str) : len));
 
   return soa_set_sdp(ss, soa_capability_sdp_kind, sdp, str, len);
 }
@@ -818,7 +818,7 @@ int soa_set_user_sdp(soa_session_t *ss,
 {
   SU_DEBUG_9(("soa_set_user_sdp(%s::%p, %p, %p, %d) called\n",
 	      ss ? ss->ss_actions->soa_name : "", ss, sdp, str, 
-	      str && len == -1 ? strlen(str) : len));
+	      str && len == -1 ? (int)strlen(str) : len));
   return soa_set_sdp(ss, soa_user_sdp_kind, sdp, str, len);
 }
 
@@ -884,7 +884,7 @@ int soa_set_remote_sdp(soa_session_t *ss,
 {
   SU_DEBUG_9(("soa_set_remote_sdp(%s::%p, %p, %p, %d) called\n",
 	      ss ? ss->ss_actions->soa_name : "", ss, sdp, str, 
-	      str && len == -1 ? strlen(str) : len));
+	      str && len == -1 ? (int)strlen(str) : len));
   return soa_set_sdp(ss, soa_remote_sdp_kind, sdp, str, len);
 }
